// Microproject/Management/model/Staff.ts
export class Staff{
    staffName: string="";
    staffNo: number=0;
    staffAdd: string="";
    staffQuali: string="";
    staffSub: string="";
    staffJoin: string="";
    staffMail: string="";
    staffId: string="";
    staffBirth: string="";
    staffExperience: number=0;
    staffCertificate: string="";
    staffSalary: number=0;
}